Split off benchmarks in separate directory.
