+++ benchmarks/audi.cpp
@@ -26,20 +26,27 @@ You should have received copies of the GNU General Public License and the
 GNU Lesser General Public License along with the Piranha library.  If not,
 see https://www.gnu.org/licenses/. */
 
-#define BOOST_TEST_MODULE monagan4_test
+#define BOOST_TEST_MODULE audi_test
 #include <boost/test/included/unit_test.hpp>
 
 #include <boost/lexical_cast.hpp>
 
 #include <piranha/init.hpp>
 #include <piranha/kronecker_monomial.hpp>
-#include <piranha/mp_integer.hpp>
+#include <piranha/polynomial.hpp>
+#include <piranha/pow.hpp>
 #include <piranha/settings.hpp>
-#include "monagan.hpp"
+
+#include "simple_timer.hpp"
 
 using namespace piranha;
 
-BOOST_AUTO_TEST_CASE(monagan4_test)
+// A performance test for truncated polynomial multiplication, in the spirit of automatic differentiation.
+// Compute:
+// (1+x1+x2+x3+x4+x5+x6+x7+x8+x9+x10)**10 * (1-x1-x2-x3-x4-x5-x6-x7-x8-x9-x10)**10
+// truncated to the total degree of 10.
+
+BOOST_AUTO_TEST_CASE(audi_test)
 {
     init();
     settings::set_thread_binding(true);
@@ -47,5 +54,15 @@ BOOST_AUTO_TEST_CASE(monagan4_test)
         settings::set_n_threads(
             boost::lexical_cast<unsigned>(boost::unit_test::framework::master_test_suite().argv[1u]));
     }
-    BOOST_CHECK_EQUAL((monagan4<integer, kronecker_monomial<>>().size()), 135751u);
+    using p_type = polynomial<double, k_monomial>;
+    p_type::set_auto_truncate_degree(10);
+    p_type x1{"x1"}, x2{"x2"}, x3{"x3"}, x4{"x4"}, x5{"x5"}, x6{"x6"}, x7{"x7"}, x8{"x8"}, x9{"x9"}, x10{"x10"};
+    auto f = math::pow(1 + x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8 + x9 + x10, 10);
+    auto g = math::pow(1 - x1 - x2 - x3 - x4 - x5 - x6 - x7 - x8 - x9 - x10, 10);
+    p_type h;
+    {
+        simple_timer t;
+        h = f * g;
+    }
+    BOOST_CHECK_EQUAL(h.size(), 122464u);
 }
