+++ benchmarks/fateman1.cpp
