+++ benchmarks/fateman1_dynamic.cpp
