+++ benchmarks/fateman1_rational.cpp
