+++ benchmarks/fateman1_unpacked_truncation.cpp
