+++ benchmarks/fateman2.cpp
