+++ benchmarks/gastineau3.cpp
