+++ benchmarks/gastineau4.cpp
