+++ benchmarks/memory_perf.cpp
@@ -43,6 +43,7 @@ see https://www.gnu.org/licenses/. */
 #include <piranha/init.hpp>
 #include <piranha/mp_integer.hpp>
 #include <piranha/settings.hpp>
+
 #include "simple_timer.hpp"
 
 using namespace piranha;
