+++ benchmarks/monagan1.cpp
@@ -26,27 +26,21 @@ You should have received copies of the GNU General Public License and the
 GNU Lesser General Public License along with the Piranha library.  If not,
 see https://www.gnu.org/licenses/. */
 
-#include <piranha/polynomial.hpp>
-
-#define BOOST_TEST_MODULE symengine_expand2b_test
+#define BOOST_TEST_MODULE monagan1_test
 #include <boost/test/included/unit_test.hpp>
 
 #include <boost/lexical_cast.hpp>
 
 #include <piranha/init.hpp>
 #include <piranha/kronecker_monomial.hpp>
-#include <piranha/math.hpp>
 #include <piranha/mp_integer.hpp>
 #include <piranha/settings.hpp>
-#include "simple_timer.hpp"
+
+#include "monagan.hpp"
 
 using namespace piranha;
-using math::pow;
 
-// Test taken from:
-// https://github.com/sympy/symengine/blob/master/benchmarks/expand2b.cpp
-
-BOOST_AUTO_TEST_CASE(symengine_expand2b_test)
+BOOST_AUTO_TEST_CASE(monagan1_test)
 {
     init();
     settings::set_thread_binding(true);
@@ -54,12 +48,5 @@ BOOST_AUTO_TEST_CASE(symengine_expand2b_test)
         settings::set_n_threads(
             boost::lexical_cast<unsigned>(boost::unit_test::framework::master_test_suite().argv[1u]));
     }
-    using p_type = polynomial<integer, k_monomial>;
-    auto x = p_type{"x"}, y = p_type{"y"}, z = p_type{"z"}, w = p_type{"w"};
-    auto f = pow(x + y + z + w, 15);
-    auto g = f + w;
-    {
-        simple_timer t;
-        BOOST_CHECK_EQUAL((f * g).size(), 6272u);
-    }
+    BOOST_CHECK_EQUAL((monagan1<integer, kronecker_monomial<>>().size()), 12341u);
 }
