+++ benchmarks/monagan4.cpp
@@ -26,34 +26,27 @@ You should have received copies of the GNU General Public License and the
 GNU Lesser General Public License along with the Piranha library.  If not,
 see https://www.gnu.org/licenses/. */
 
-#ifndef PIRANHA_GASTINEAU3_HPP
-#define PIRANHA_GASTINEAU3_HPP
+#define BOOST_TEST_MODULE monagan4_test
+#include <boost/test/included/unit_test.hpp>
 
-#include <piranha/polynomial.hpp>
-#include "simple_timer.hpp"
+#include <boost/lexical_cast.hpp>
 
-namespace piranha
+#include <piranha/init.hpp>
+#include <piranha/kronecker_monomial.hpp>
+#include <piranha/mp_integer.hpp>
+#include <piranha/settings.hpp>
+
+#include "monagan.hpp"
+
+using namespace piranha;
+
+BOOST_AUTO_TEST_CASE(monagan4_test)
 {
-
-template <typename Cf, typename Key>
-inline polynomial<Cf, Key> gastineau3()
-{
-    typedef polynomial<Cf, Key> p_type;
-    p_type u("u"), v("v"), w("w"), x("x"), y("y");
-
-    auto f = (1 + u * u + v + w * w + x - y * y);
-    auto g = (1 + u + v * v + w + x * x + y * y * y);
-    auto tmp_f(f), tmp_g(g);
-    for (int i = 1; i < 28; ++i) {
-        f *= tmp_f;
-        g *= tmp_g;
+    init();
+    settings::set_thread_binding(true);
+    if (boost::unit_test::framework::master_test_suite().argc > 1) {
+        settings::set_n_threads(
+            boost::lexical_cast<unsigned>(boost::unit_test::framework::master_test_suite().argv[1u]));
     }
-    g += 1;
-    {
-        simple_timer t;
-        return f * g;
-    }
+    BOOST_CHECK_EQUAL((monagan4<integer, kronecker_monomial<>>().size()), 135751u);
 }
-}
-
-#endif
