+++ benchmarks/pearce1.cpp
