+++ benchmarks/pearce1_dynamic.cpp
