+++ benchmarks/pearce1_rational.cpp
