+++ benchmarks/pearce2.cpp
