+++ benchmarks/pearce2.hpp
@@ -26,17 +26,18 @@ You should have received copies of the GNU General Public License and the
 GNU Lesser General Public License along with the Piranha library.  If not,
 see https://www.gnu.org/licenses/. */
 
-#ifndef PIRANHA_GASTINEAU2_HPP
-#define PIRANHA_GASTINEAU2_HPP
+#ifndef PIRANHA_PEARCE2_HPP
+#define PIRANHA_PEARCE2_HPP
 
 #include <piranha/polynomial.hpp>
+
 #include "simple_timer.hpp"
 
 namespace piranha
 {
 
 template <typename Cf, typename Key>
-inline polynomial<Cf, Key> gastineau2()
+inline polynomial<Cf, Key> pearce2()
 {
     typedef polynomial<Cf, Key> p_type;
     p_type x("x"), y("y"), z("z"), t("t"), u("u");
@@ -45,7 +46,7 @@ inline polynomial<Cf, Key> gastineau2()
     auto tmp_f(f);
     auto g = (u + t + z * z * 2 + y * y * y * 3 + x * x * x * x * x * 5 + 1);
     auto tmp_g(g);
-    for (int i = 1; i < 25; ++i) {
+    for (int i = 1; i < 16; ++i) {
         f *= tmp_f;
         g *= tmp_g;
     }
