+++ benchmarks/simple_timer.hpp
