+++ benchmarks/symengine_expand2b.cpp
@@ -26,33 +26,41 @@ You should have received copies of the GNU General Public License and the
 GNU Lesser General Public License along with the Piranha library.  If not,
 see https://www.gnu.org/licenses/. */
 
-#ifndef PIRANHA_FATEMAN1_HPP
-#define PIRANHA_FATEMAN1_HPP
-
 #include <piranha/polynomial.hpp>
+
+#define BOOST_TEST_MODULE symengine_expand2b_test
+#include <boost/test/included/unit_test.hpp>
+
+#include <boost/lexical_cast.hpp>
+
+#include <piranha/init.hpp>
+#include <piranha/kronecker_monomial.hpp>
+#include <piranha/math.hpp>
+#include <piranha/mp_integer.hpp>
+#include <piranha/settings.hpp>
+
 #include "simple_timer.hpp"
 
-namespace piranha
-{
+using namespace piranha;
+using math::pow;
 
-template <typename Cf, typename Key>
-inline polynomial<Cf, Key> fateman1(unsigned long long factor = 1u)
+// Test taken from:
+// https://github.com/sympy/symengine/blob/master/benchmarks/expand2b.cpp
+
+BOOST_AUTO_TEST_CASE(symengine_expand2b_test)
 {
-    typedef polynomial<Cf, Key> p_type;
-    p_type x("x"), y("y"), z("z"), t("t");
-    auto f = x + y + z + t + 1;
-    auto tmp(f);
-    for (auto i = 1; i < 20; ++i) {
-        f *= tmp;
+    init();
+    settings::set_thread_binding(true);
+    if (boost::unit_test::framework::master_test_suite().argc > 1) {
+        settings::set_n_threads(
+            boost::lexical_cast<unsigned>(boost::unit_test::framework::master_test_suite().argv[1u]));
     }
-    if (factor > 1u) {
-        f *= factor;
-    }
+    using p_type = polynomial<integer, k_monomial>;
+    auto x = p_type{"x"}, y = p_type{"y"}, z = p_type{"z"}, w = p_type{"w"};
+    auto f = pow(x + y + z + w, 15);
+    auto g = f + w;
     {
         simple_timer t;
-        return f * (f + 1);
+        BOOST_CHECK_EQUAL((f * g).size(), 6272u);
     }
 }
-}
-
-#endif
