+++ tests/CMakeLists.txt
@@ -1,21 +1,5 @@
 # Setup for the split compilation of tests.
 
-# This is just a simple counter variable, internal use only.
-set(_PIRANHA_TEST_NUM "0")
-
-# Check splitting options. These need to be set from the command line.
-# - PIRANHA_TEST_NSPLIT: number of chunks into which the unit tests will be divided (must be > 1).
-# - PIRANHA_TEST_SPLIT_NUM: 0-based index of the chunk to run.
-if(PIRANHA_TEST_NSPLIT AND "${PIRANHA_TEST_SPLIT_NUM}" STREQUAL "")
-	message(FATAL_ERROR "Test splitting was requested, but the PIRANHA_TEST_SPLIT_NUM variable was not set.")
-elseif(NOT PIRANHA_TEST_NSPLIT AND NOT "${PIRANHA_TEST_SPLIT_NUM}" STREQUAL "")
-	message(FATAL_ERROR "The PIRANHA_TEST_SPLIT_NUM variable was set, but test splitting was not requested.")
-endif()
-
-if(PIRANHA_TEST_NSPLIT)
-	message(STATUS "Tests will be split into ${PIRANHA_TEST_NSPLIT} chunks. The chunk with index ${PIRANHA_TEST_SPLIT_NUM} will be processed.")
-endif()
-
 function(ADD_PIRANHA_TESTCASE arg1)
 	if(PIRANHA_TEST_NSPLIT)
 		math(EXPR _PIRANHA_TEST_NUM "(${_PIRANHA_TEST_NUM} + 1) % ${PIRANHA_TEST_NSPLIT}")
@@ -28,8 +12,6 @@ function(ADD_PIRANHA_TESTCASE arg1)
 			# option is selected.
 			target_compile_definitions(${arg1} PRIVATE PIRANHA_RUN_ON_VALGRIND)
 		endif()
-		# Pass the current source dir as a definition, so we can load data for tests from the source dir.
-		target_compile_definitions(${arg1} PRIVATE PIRANHA_TESTS_DIRECTORY="${CMAKE_CURRENT_SOURCE_DIR}")
 		target_link_libraries(${arg1} piranha Boost::system Boost::filesystem)
 		set_property(TARGET ${arg1} PROPERTY CXX_STANDARD 11)
     set_property(TARGET ${arg1} PROPERTY CXX_STANDARD_REQUIRED YES)
@@ -37,32 +19,6 @@ function(ADD_PIRANHA_TESTCASE arg1)
 		add_test("${arg1}" "${arg1}")
 	endif()
 endfunction()
-
-# MACRO(ADD_PIRANHA_TESTCASE arg1)
-# 	IF(CMAKE_BUILD_TYPE STREQUAL "Debug")
-# 		if(PIRANHA_TEST_NSPLIT)
-# 			math(EXPR _PIRANHA_TEST_NUM "(${_PIRANHA_TEST_NUM} + 1) % ${PIRANHA_TEST_NSPLIT}")
-# 		endif()
-# 		if(NOT PIRANHA_TEST_NSPLIT OR "${PIRANHA_TEST_SPLIT_NUM}" STREQUAL "${_PIRANHA_TEST_NUM}")
-# 			ADD_EXECUTABLE("${arg1}" "${arg1}.cpp")
-# 			TARGET_LINK_LIBRARIES("${arg1}" ${MANDATORY_LIBRARIES} ${Boost_SYSTEM_LIBRARY} ${Boost_FILESYSTEM_LIBRARY})
-# 			ADD_TEST("${arg1}" "${arg1}")
-# 		endif()
-# 	ENDIF()
-# ENDMACRO()
-
-# MACRO(ADD_PIRANHA_PERFORMANCE_TESTCASE arg1)
-# 	IF(CMAKE_BUILD_TYPE STREQUAL "Release")
-# 		if(PIRANHA_TEST_NSPLIT)
-# 			math(EXPR _PIRANHA_TEST_NUM "(${_PIRANHA_TEST_NUM} + 1) % ${PIRANHA_TEST_NSPLIT}")
-# 		endif()
-# 		if(NOT PIRANHA_TEST_NSPLIT OR "${PIRANHA_TEST_SPLIT_NUM}" STREQUAL "${_PIRANHA_TEST_NUM}")
-# 			ADD_EXECUTABLE("${arg1}_perf" "${arg1}_perf.cpp")
-# 			TARGET_LINK_LIBRARIES("${arg1}_perf" ${MANDATORY_LIBRARIES} ${Boost_SYSTEM_LIBRARY} ${Boost_FILESYSTEM_LIBRARY})
-# 			ADD_TEST("${arg1}_perf" "${arg1}_perf")
-# 		ENDIF()
-# 	ENDIF(CMAKE_BUILD_TYPE STREQUAL "Release")
-# ENDMACRO(ADD_PIRANHA_PERFORMANCE_TESTCASE)
 
 ADD_PIRANHA_TESTCASE(array_key)
 ADD_PIRANHA_TESTCASE(atomic_utils)
@@ -151,36 +107,3 @@ ADD_PIRANHA_TESTCASE(trigonometric_series)
 ADD_PIRANHA_TESTCASE(tuning)
 ADD_PIRANHA_TESTCASE(type_traits)
 ADD_PIRANHA_TESTCASE(ulshift)
-
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(audi)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(estimation)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(evaluate)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(fateman1)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(fateman1_dynamic)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(fateman1_rational)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(fateman1_unpacked)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(fateman1_unpacked_truncation)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(fateman2)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(gastineau1)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(gastineau2)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(gastineau3)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(gastineau4)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(memory)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(monagan1)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(monagan2)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(monagan3)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(monagan4)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(monagan5)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(power_series)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(pearce1)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(pearce1_dynamic)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(pearce1_rational)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(pearce1_unpacked)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(pearce2)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(pearce2_unpacked)
-# if(PIRANHA_WITH_MSGPACK AND PIRANHA_WITH_BZIP2)
-# 	ADD_PIRANHA_PERFORMANCE_TESTCASE(perminov1)
-# endif()
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(rectangular)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(s11n)
-# ADD_PIRANHA_PERFORMANCE_TESTCASE(symengine_expand2b)
